search_app/providers/universal.py
from typing import List, Dict
import aiohttp
from .base import BaseSearchProvider
from django.conf import settings

class UniversalSearchProvider(BaseSearchProvider):
    """
    AI-gesteuerter universeller Suchprovider.
    Nutzt LLMs für Queryanalyse und Ergebnisverarbeitung.
    """
    
    def __init__(self, **kwargs):
        super().__init__(**kwargs)
        self.llm_url = settings.LLM_API_URL
        self.embedding_url = settings.EMBEDDING_API_URL
        self.searxng_url = settings.SEARXNG_URL

    async def analyze_query(self, query: str) -> List[Dict]:
        """
        Analysiert die Query mit LLM um relevante Provider zu identifizieren
        """
        prompt = f"""
        Analyze this search query and determine which search providers would be most relevant.
        Query: {query}
        
        Consider these aspects:
        1. Query type (factual, academic, technical, etc.)
        2. Required data sources
        3. Time sensitivity
        4. Domain specificity
        
        Return a list of recommended providers and their priorities.
        """
        
        async with aiohttp.ClientSession() as session:
            async with session.post(self.llm_url, json={'prompt': prompt}) as response:
                analysis = await response.json()
                
        return self._parse_provider_recommendations(analysis['response'])

    async def search(self, query: str, filters: Dict = None) -> List[Dict]:
        """
        Führt die universelle Suche durch
        """
        # 1. Get initial results from SearXNG
        raw_results = await self._fetch_searxng_results(query)
        
        # 2. Generate embeddings for query and results
        embeddings = await self._generate_embeddings(query, raw_results)
        
        # 3. Re-rank results using embeddings
        ranked_results = self._rerank_results(raw_results, embeddings)
        
        # 4. Enhance results with AI-generated insights
        enhanced_results = await self._enhance_results(query, ranked_results)
        
        return enhanced_results

    async def _fetch_searxng_results(self, query: str) -> List[Dict]:
        """Fetches results from SearXNG"""
        async with aiohttp.ClientSession() as session:
            async with session.get(
                f"{self.searxng_url}/search", 
                params={'q': query, 'format': 'json'}
            ) as response:
                return await response.json()

    async def _generate_embeddings(self, query: str, results: List[Dict]) -> Dict:
        """Generates embeddings for query and results"""
        texts = [query] + [r['snippet'] for r in results]
        
        async with aiohttp.ClientSession() as session:
            async with session.post(
                self.embedding_url, 
                json={'texts': texts}
            ) as response:
                return await response.json()

    def _rerank_results(self, results: List[Dict], embeddings: Dict) -> List[Dict]:
        """Re-ranks results based on embedding similarity"""
        # Implementierung von Similarity Scoring
        return sorted(results, key=lambda x: x.get('relevance_score', 0), reverse=True)

    async def _enhance_results(self, query: str, results: List[Dict]) -> List[Dict]:
        """Enhances results with AI-generated insights"""
        # Implementierung von AI Enhancement
        return results 