requirements.txt
# Python >= 3.9
# CUDA >= 11.8 recommended for GPU support
# System requirements:
# - Tesseract OCR for pytesseract
# - FFmpeg for audio processing
# - Git LFS for large model files

# Core Django & Web Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-filter==23.3
drf-yasg==1.21.7
channels==4.0.0
channels-redis==4.1.0
django-guardian==2.4.0
djangorestframework-simplejwt==5.3.0
whitenoise==6.5.0
daphne==4.0.0
gunicorn==21.2.0

# Database & Caching
psycopg2-binary==2.9.9
django-redis==5.4.0
redis==5.0.1

# Security & Configuration
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
cryptography==41.0.5
django-environ==0.11.2

# AI/ML Ecosystem
## OpenAI & Other LLM Providers
openai==1.3.5  # Requires API key
anthropic==0.5.0  # Requires API key
groq==0.4.0  # Requires API key
google-generativeai==0.8.4  # Requires API key
ollama==0.4.7

## Mistral AI
mistral-common[opencv]==1.4.1  # Requires API key
mistral-inference==1.5.0

## PyTorch Ecosystem (CUDA support)
# For CUDA 11.8 (GPU support): +cu118
torch==2.2.1
torchvision==0.17.1
torchaudio==2.2.1
# For CPU only, remove +cu118 suffix

# PyTorch Extensions
# torch-scatter==2.1.1  # Requires matching CUDA version
# torch-sparse==0.6.17  # Requires matching CUDA version
# torch-geometric==2.4.0  # Requires matching CUDA version
einops==0.7.0

## Hugging Face Ecosystem
transformers[vision]==4.35.2  # Includes vision-specific dependencies
huggingface_hub==0.19.4  # Requires authentication for private models
accelerate==0.24.1
bitsandbytes==0.41.1  # For model quantization
hf_transfer==0.1.4
vllm==0.6.1

# Vision & Image Processing
Pillow==10.3.0
opencv-python==4.8.1.78  # Use opencv-python-headless for servers
albumentations==1.3.1
segment-anything-py==1.0  # Updated package name
timm==0.9.7
byaldi==0.0.7  
qwen-vl-utils==0.0.10
qwen==0.1.1
# qwen-vl-chat==1.0.3  # Load Qwen‑VL‑Chat directly via Transformers in your code:
colpali-engine==0.1.0  # Updated package name
layoutparser==0.3.4  # Requires additional model weights

# Data Processing & Analysis
numpy==1.26.0
pandas==2.1.1
scipy==1.11.4
spacy==3.7.2  # After install, run: python -m spacy download en_core_web_sm
nltk==3.8.1  # Requires: done
bertopic==0.15.0
rank-bm25==0.2.2
urlextract==1.8.0
beautifulsoup4==4.12.2
readability-lxml==0.8.1
cachetools==5.3.2
flake8==7.1.2

# Document Processing
PyPDF2==3.0.1
pdf2image==1.16.3  # Requires poppler installation => should be installed
pdfminer.six==20221105
python-magic==0.4.27
pytesseract==0.3.10  # Requires Tesseract OCR installation => done
docx2pdf==0.1.8  # Requires Microsoft Word or LibreOffice
markdown==3.5.1
openpyxl==3.1.2
python-pptx==0.6.21
PyMuPDF==1.23.7
python-docx  # for reading/writing .docx files
pypdf2  # for PDF operations
pdf2docx  # for PDF to DOCX conversion

# Audio Processing
faster-whisper==0.10.0  # Updated from whisper-fast
##distil-whisper==0.1.0 => model_id = "distil-whisper/distil-large-v3" usage
librosa==0.10.1 
soundfile==0.12.1
audioread==3.0.1
pydub==0.25.1 
noisereduce==3.0.0
pyannote.audio==3.1.1  # Requires authentication and model license
torch-audiomentations==0.11.0
pyannote.core==5.0.0
ffmpeg-python==0.2.0
imageio-ffmpeg==0.6.0

# Development & Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.0
pre-commit==3.5.0
jupyter==1.0.0
ipython==8.17.2
coverage==7.3.2

# Task Queue & Monitoring
celery==5.3.4
flower==2.0.1
prometheus-client==0.17.1
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-exporter-otlp==1.20.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.7
plantuml==0.3.0

# Utilities
requests==2.31.0
tqdm==4.66.1
tomli==2.0.1
pkgconfig==1.5.5
python-Levenshtein==0.27.1
cmake==3.31.6
psutil==5.9.6

# Optional: Platform-specific requirements
# Windows-only packages:
# pywin32==306  # Required for Windows service support and system integration
#               # Used for Windows-specific features like service management
#               # and system tray integration
# 
# Linux-only packages:
# python-magic-bin==0.4.14  # Required for file type detection on Linux

# Post-installation steps:
# 1. Install system dependencies:
#    - FFmpeg
#    - Tesseract OCR
#    - Git LFS
#    - poppler-utils
# 2. Download language models:
#    python -m spacy download en_core_web_sm
#    python -m nltk.downloader popular
# 3. Set up API keys for:
#    - OpenAI
#    - Anthropic
#    - Google AI
#    - Mistral AI
#    - HuggingFace
# 4. Download required model weights for:
#    - segment-anything
#    - layoutparser
#    - pyannote.audio